Вынесен тест абстрактных методов в отдельный файл
